+++ NowoHotReloadBundle.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Nowo\HotReloadBundle;
 
+use Nowo\HotReloadBundle\DependencyInjection\Compiler\TwigPathsPass;
 use Nowo\HotReloadBundle\DependencyInjection\HotReloadExtension;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -16,6 +18,12 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class NowoHotReloadBundle extends Bundle
 {
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+        $container->addCompilerPass(new TwigPathsPass());
+    }
+
     public function getContainerExtension(): ?ExtensionInterface
     {
         if ($this->extension === null) {